Build both ostree and regular images from the same manifests
This adds a new variable `image_type` which defaults to `"ostree"`, but
we can override it with  `-D image_type="\"regular\""` to make it build
a regular image.

This works by using $image_type in the includes so that we get the right
type.

To allow per-image-type packages this also introduces an image_rpms
variable which is used by the main recipe.