Add new directboot image type
This is similar to the regular image type, except it doesn't boot
via grub but via direct uefi boot. To do this it reuses the
ostree-tree pipeline (to avoid needing more includes from other manifests)
to build the kernel and initrd, and then the image manifest copies
the kernel, etc directly into the EFI boot partition.

For the boot to work we include in the EFI a boot.csv file that lets
shim create a proper boot entry with an uefi argument so that the kernel
can find the initrd.

The goal of this is two-fold:
 1) Abstract out how the kernel gets created, allowing for easy
    switching in of 3rd party kernels.
 2) Remove grub from the boot path for faster boots.

This requires mpp-if and the gunzip stage from osbuild.