cs9: Add the automotive custom repo from centos cbs
This adds the repos at:
 https://buildlogs.centos.org/9-stream/automotive/$arch
to the images.

This is the `automotive9s-packages-main-testing` tag from cbs:
  https://cbs.centos.org/koji/taginfo?tagID=2461

Later we may want to change it to the
`automotive9s-packages-main-release` tag, but that is locked until the
mirror situation is cleared up.

In addtition this adds the cs9_crb_repos and cs9_debug_repos, which
make it easier to add the coder-ready-builder and the debuginfo repos
to an image (for example when doing development/debugging).