cs9: Use the kernel-auto rpm instead of the regular kernel.
Note that this doesn't (currently) install the kernel-auto package,
just the kernel-auto-core one, because the kernel-auto package adds a
dependency on auto-setup, which we don't currently have. This is a
conversion issue from the kernel-rt package which has a corresponding
rt-setup package, but that is also not available in the automotive repos.
Work is ongoing to clean this up.