Drop unnecessary bootloader packages
The grub2-pc and grub2-pc-modules packages are only needed for legacy BIOS
x86 platforms. We only support x86 UEFI platforms, so these can be dropped.

There are also other packages that are not needed, the grub2-efi-ia32 and
shim-ia32 packages are to allow booting a x86_64 OS on a machine that has
a 32-bit UEFI firmware. But platforms like these are also not supported.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>