Makefile: Support checkpoints with osbuildvm
Pass along all --checkpoint args to osbiuld in vm, and for all checkpoints
that were written, tar them up and send out as output.

Then when starting osbuild, extract the info about what checkpoints are
needed and if we have any stored tar:ed checkpoints, add them to the vm work
image.

The tar:ed up cached refs are stored in the refs_tars subdir in the
osbuild store, so we also add this to make clean_caches.