Add a "container" output type to our image building ipp
This allows to build a container image using osbuild and our
standard tooling.
The container is provided as a tar file which can then be easily
imported into podman using: `podman import -i <file>.tar`.
You can then find it in `podman images` and run it as any other
container.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>