containers: Use /usr/share/containers/storage for container store
The preloaded container images are stored in the read-only part of the
image. This is a better fix of how you would use most images in an
automotive situation.

It also fits better with shipping/updating the container image with
ostree. The regular storage is in /var which is not under ostree
control, so updates to this in the image wouldn't necessarily be
picked up in an ostree update.