containers: Add and use custom nginx container for demo
There was some issues with the ubi based nginx which made it not work
with the ostree based image (it had whiteouts in some container
layer), so lets do our own demo containers.

This way we can be in better control of what is in the containers, and
change with what to demo easier.