Simplify Makefile in favour of runosbuild shellscript
This moves some of the work that the Makefile did to run osbuild into
a shellscript. This simplifies things, because some of the stuff
this was doing isn't a great fit for Makefiles.

Additionally this lays the ground for some additional work when running
osbuild that would have made the Makefile even more complex.