Makefile: Add support for OSTREE_REPO for parent-commit handling
If OSTREE_REPO is set in the make command, then this repo is inspected
and the commit id for all the current refs in it is passed to osbuild-mpp,
allowing it to pick up the right parent commit in the new ostree commit.

This allows a produced ostree commit to continue the history of an
existing repo.