ostree: Don't update grub config each deploy
The default bootloader option is auto, which causes ostree to rewrite
the grub config based on the current config each time. This replaces
the custom osbuild grub config that we installed and seems unsafe wrt
to things like rollback guaranteed to work, so this disables that
by setting bootloader=none.