Add support for luks support and auto-unlock
If use_luks is defined, the root filesystem is encrypted using the
passphrase from the `luks_passphrase` variable (default is
`password`).

If luks_auto_unlock is true, then the passphrase is stored in
/usr/.auto-unlock-key and is added to the initrd. Then the right
kernel cmdline option is added to auto-unlock the rootfs using this
key.

NOTE: For some reason this does not currently work with the direct
image type.