Add a gadget image
This image is based on the developer image, so it will have most
of the packages available in that image.
In addition that image pulls in the pi_gadget package which configure
the raspberry pi as an USB gadget. It also pull in cockpit and cockpit
podman. So the image opens up the port 9090 to make the cockpit UI
accessible and enable/start cockpit.socket.
With all this, you can build a gadget image, flash it on the SD card,
boot it on your raspberry pi.
Once it's booted, you can access it over USB (as an USB gadget) and access
cockpit's UI at http://<ip>:9090 (where IP is 192.168.42.15 when
accessed via the USB gadget).

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>