runvm: add the possibility to pass a compiled QEMU binary
For some developer who is testing any image or any QEMU feature like
TCG. It would be nice to have the ability to set your own QEMU path to
test it. Some QEMU docs refer it to QEMU environment variable which is
the binary path. But here, we are proposing just the directory to follow
the script architecture logic. Developers can still use a local install
if needed. The commit removes qemu-kvm binary because it is a link for
the qemu binary which reflects the same platform of the host.

Signed-off-by: Julio Faracco <jfaracco@redhat.com>