Add aboot suport and abootqemu target (for regular only atm)
This adds support for generating aboot images and paritions, based
on abootimg and aboot-update from auto-sig repo.

This also requires the build host to have osbuild-aboot installed
from the  `alexl/osbuild-aboot` COPR. This has some new experimental
osbuild stages.

The way this works is:
 * abootqemu target adds some rpms to support qemu w/ dtb
 * abootqemu target enables use_aboot and use_abootpart
 * if use_aboot is true, the common code does:
   * Adds the required rpms
   * creates a /boot/aboot.cfg file
   * runs aboot-update based on it which creates a /boot/aboot-$version.img file
 * if aboot_usepart is true, image does:
   * add aboot partition to partition table
   * copies the generated aboot image to the aboot partition

aboot requires a devicetree, and for abootqemu we use one generated by
qemu in the autosig-qemu-dtb rpm. QEmu also needs a custom firmware to
boot with aboot. This is available in the autosig-u-boot rpm (which is
/boot/u-boot.img in the built image).