Correctly compute dependent variables
We can't use mpp-if in the defaults include snippet, becase by
that time we have not read the image-*, target-* and arch-* snippets
so not all variables are set yet. We fix this by adding a final
compute-vars include file at the end when everything is set.

This also moves all conditional stuff from defaults to compute-vars.