Add kernel_package and kernel_version variables
Allow a user to specify separately the specific RPM version for the
kernel package. This is a useful feature when a user wants to install
and use an older kenel rpm despite the repositories having newer
versions.

The original kernel_rpm variable is kept for backwards compatibility and
will override the effects of kernel_package/kernel_version.