regular: Move dracut option to config file
Instead of passing all the image options to the dracut command during
the image build we set a config file and then run it without arguments.

The advantage is that later runs of dracut on the image (e.g. if updating
kernel) will reuse the same options.

I verified that this generates the same initial initrd.