Added rules for merge-request
For MRs in which the files can affect the ci pipeline have changed, it should trigger the pipeline.
For MRs in which only the package-list was changed, use the DISABLE flag to not trigger the creation of non-sample-images in p-a-c.
Except for the files are mentioned in the 2 new rules, any other change shouldn't change the success of the ci pipeline, thus shouldn't trigger it.