Set ostree sysroot read-only as false
After osbuild fixed a bug with the ostree config stage
(https://github.com/osbuild/osbuild/pull/1129), the ostree images
started to fail booting.

It seems like the, althought the ostree sysroot was set as true,
it wasn't enabled as such. Now that osbuild is actually using that
setting, the sysroot is set as read-only, but it's causing the system
to be unbooteable.

The desired config is the sysroot to be read-only, but it'll be set
as false to workaround that issue, until we're able to find and fix
the root cause.