container image: Use transient image store
Our images pre-load container images into /usr/share/containers/storage,
so /var/lib/containers/storage is mostly empty. In addtition we never
reuse containers, always creating new ones each run, and passing `--rm`
to auto-delete them. This makes this storage directory mostly a place
that keeps track of what containers are running atm.

This change installs the containers-transient-store which makes the
storage directory a tmpfs mountpoint, thus making all the container
state transient (i.e. not persist across reboots).

There are two advantages to this:

 * Tmpfs is much faster, especially the fsync() calls that podman does
 * Each boot is guaranteed to be fresh, there is not leftover container
   runtime state across reboots.

However, the default directory for storing the backing state of
volumes is /var/lib/containers/storage/volumes, and we don't want
that to be transient, so we also have a custom containers.conf that
points this to /var/lib/containers/volumes.