osbuild-manifests: drop rngd-tools and rngd.service
Drop by default the rng-tools package from the sample-images. This is
because in the CI/CD for sample-images, we're seeing rngd.service fail
on aarch64 AWS EC2 VMs used to smoke test the aarch64 images as of the
new version of the package (v6.15-2).

The issue being seen relates to a change in rng-tools v6.15-2, where a
timeout was added by default to the "jitter" entropy source which
caused rngd to crash if it could not gather enough entropy fast enough
from said source. The debugging logs on the ARM64 AWS EC2 VM showed that
there was no other hardware rng device available and so rngd was
dependent on the jitter source on these machines. Unfortunately the jitter
source was also slow to seed on these machines and requires a
significantly longer time than the default to gather enough entropy.

Increasing the jitter timeout to 60s seemed sufficient to workaround the
error. Removing rng-tools from the tested images also seemed to show no
issue upfront. So this commit takes the path of removing rng-tools from
the defaults in sample-images.

Users who need rng-tools to help seed entropy on embedded devices which
might need certain configuration of the rng devices can add it back in
with their custom mpp files.