images: Clean up container support and use new config options
This adds a new option use_containers which when set (and its now set in our
container images) automatically creates storage.conf and containers.conf,
as well as adding the new podman-snapshot dnf repo.

Also, with the snapshot we don't need to create /etc/containers/storage,
so drop that. We also now set transient_store and read_only by default.

Signed-off-by: Alexander Larsson <alexl@redhat.com>