Add containersperf image
This contains the latest podman snapshot (via use_containers=true) and
sets it up for transient store. It also contains hyperfine and some
basic tools to make it easy to do profiling/testing, as well as dnf so
that the image can be extended later.

It also adds the rcupdate.rcu_normal_after_boot=1 kernel option, which is
the default, so this doesn't affect anything but makes it easy to change
it to 0 to debug RT-vs-RCU slowness issues.