Use absolute paths for qemu dtb that is in a separate rpm
Normally dtbs are part of the kernel rpm which means the dtbs are
installed to specific dirs with versioning info included, examples
below:

/boot/dtb-6.2.3-450.1.asahi.fc37.aarch64/qcom/sdm845-oneplus-enchilada.dtb
/usr/lib/modules/6.2.3-450.1.asahi.fc37.aarch64/dtb/qcom/sdm845-oneplus-enchilada.dtb

It is difficult for .dtb's kept in other rpms to align with this
approach because of the need to move the dtb based on the kernel you are
currently on. For this reason, if we use an absolute path, grab the dtb
there, otherwise search in the relevant /boot dir or /usr/lib/modules
dir. This way works for both dtb's packaged in a kernel rpm or in a
separate rpm.