Sign and verify composefs image
 * During build we create a build-specific key/cert pair
 * During commit, the new rpm-ostree will automatically create a
   composefs digest in the commit metadata.
 * After commit we sign the commit with the private key.
 * The public cert is added as /etc/ostree/commit.key in the commit
 * The cert is also added to the initrd
 * We add ot-composefs=signed=/etc/ostree/commit.key to the kernel
 cmdline

During boot, in the initrd, ostree-prepare-boot will find the commit,
validate it against the public key, and then extract the composefs
digest from it, and ensure that the compoosefs file has the expected
fs-verity digest.

This extends the trust in a signed initrd to all of the userspace in
the ostree commit.