do not pass rootfs_uuid with ABL like bootloader
Add a use_abl flag to set or not the rootfs_uuid depending on the
target. The default is false as this is target specific.

Boards relying on the ABL bootloader are expected to receive the root=
parameter from the bootloader. The rootfs partition is selected when the
A/B slot is determined by the firmware then bootloader.

For the ABL case, if the root= parameter is set in the boot-image, the
bootloader will not try to set the rootfs for that slot and instead use
the one from the boot-image.

Using the images from this repository at this time, the UUID is static.
This can lead to situations where two different partitions are flashed
with the same UUID. The first one that the system finds will be used,
wether or not this is the most recently flashed or the desired one is
undetermined.

The change is limited to developper and minimal images for now as these
are the only images affected by such circumstances that I know about.
This may have to be added to more image types or factorized in a more
generic way.

Signed-off-by: Eric Chanudet <echanude@redhat.com>