omit unneeded dracut modules
Omit the i18n module since this dracut module installs a large number
of files in the initramfs. Let's also omit the qemu and virtiofs since
they aren't needed for our current use cases.

The bash module is also removed, however the bash binary will still end
up in the final initramfs since the base module brings it in.

This change reduces the initramfs for the RideSX4 target from ~18MB to
~15MB. Just about all of the space savings came from removing the i18n
module.

Signed-off-by: Brian Masney <bmasney@redhat.com>