prune lots more modules from the ridesx4 initramfs
Prune a large list of modules that aren't needed for the RideSX4 target
in an attempt to get the initramfs as small as possible. After this
change, these are the only modules that are remaining for this target
inside the initramfs:

    crypto/zstd.ko.zst
    drivers/clk/qcom/clk-qcom.ko.zst
    drivers/clk/qcom/clk-rpmh.ko.zst
    drivers/clk/qcom/gcc-sa8775p.ko.zst
    drivers/crypto/qcom-rng.ko.zst
    drivers/devfreq/governor_simpleondemand.ko.zst
    drivers/interconnect/qcom/icc-bcm-voter.ko.zst
    drivers/interconnect/qcom/icc-rpmh.ko.zst
    drivers/interconnect/qcom/qnoc-sa8775p.ko.zst
    drivers/phy/qualcomm/phy-qcom-qmp-ufs.ko.zst
    drivers/pinctrl/qcom/pinctrl-msm.ko.zst
    drivers/pinctrl/qcom/pinctrl-sa8775p.ko.zst
    drivers/regulator/qcom-rpmh-regulator.ko.zst
    drivers/soc/qcom/cmd-db.ko.zst
    drivers/soc/qcom/qcom-geni-se.ko.zst
    drivers/soc/qcom/qcom_rpmh.ko.zst
    drivers/soc/qcom/rpmhpd.ko.zst
    drivers/tty/serial/qcom_geni_serial.ko.zst
    drivers/ufs/core/ufshcd-core.ko.zst
    drivers/ufs/host/ufshcd-pltfrm.ko.zst
    drivers/ufs/host/ufs_qcom.ko.zst
    fs/ext4/ext4.ko.zst
    fs/jbd2/jbd2.ko.zst
    fs/mbcache.ko.zst
    lib/zstd/zstd_compress.ko.zst

Note that dracut has a '--drivers' to list the modules that will be
exclusively be added to the initramfs. However, osbuild's dracut stage
doesn't support this at the moment. I'll get that feature added to
osbuild, however that may take up to a few weeks before we can merge
that change here. So let's just start with this really long list right
now, and I'll update this target once everything is merged and in an
osbuild release.

Signed-off-by: Brian Masney <bmasney@redhat.com>