don't add quiet to the kernel command line when the debug kernel is used
Previously only the minimal image had the quiet kernel command line
option added to help with the overall boot speed. This is not desirable
for the images that use the debug kernel. Let's change this by looking
at the use_debug variable. This change was tested by running:

    make DEFINES+='kernel_package="kernel-automotive-debug"' \
                    cs9-ridesx4-developer-regular.aarch64.aboot

and with:

    make cs9-ridesx4-developer-regular.aarch64.aboot

Signed-off-by: Brian Masney <bmasney@redhat.com>