target/ridesx4: Use net.naming-scheme=v252
This SoC has 2 stmmac platform devices on it for networking. Without a
new enough[0] scheme these devices do not get a persistent name, which
can be troublesome for configuration purposes.

The el9 family is stuck on a distro custom scheme, rhel-9.0, which maps
to v250 by default[1]. Let's specify using v252 to ensure that we can
get a persistent name matching the devicetree alias.

[0] https://www.freedesktop.org/software/systemd/man/latest/systemd.net-naming-scheme.html#v252
[1] https://gitlab.com/redhat/centos-stream/rpms/systemd/-/blob/ea71a49292e67e7d01f1092cd12810c17a511a37/0112-Net-naming-scheme-for-RHEL-9.0.patch

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>