Boot time optimization --no-compress over --gzip on rpm-ostree images
We recently changed from --lz4 to --gzip in ostree images as rpm-ostree
currently doesn't support compression algorithm's other than gzip. But,
we are now small enough in terms of initrd, kernel, dtb and cmdline that
no compression is an option and is faster to boot than gzip. In the
longer-term it's likely we change to a faster decompression algorithm
like lz4 or zstd again.

Signed-off-by: Eric Curtin <ecurtin@redhat.com>