runosbuild: Support multiple extensions
If you use an extension like image.foo+bar runosbuild now passes
--export foo-export-name to the osbuild, and exports *both* the
image.foo and the image.bar file.

This can be used to create multiple parts of a build, which is often
useful as you want both a copy of the ostree commit, as well as the
a disk image that embedds the ostree commit.

Later comments will add supported combinations to the Makefile.