Never add / to fstab if use_ostree, to make composefs optional
If the real rootfs partition is listed in fstab then
systemd-remount-fs.service will try to remount it read-write.  This
causes problems when composefs is used, because systemd tries to
remount /, which now is the composefs mount, not the device mount. And
anyway, remount is not actually necessary for us, as we pass "rw" to
the kernel if use_ostree is set.

We fix this by just dropping / from the fstab, which means the fstab
generated in the ostree commit works both with and without composefs,
making the use of composefs something that only depends on how the
image is deployed.