target/ridesx4: use dracut's add_drivers to exclusively include kernel modules
Rather than excluding kernel modules, let's use dracut's add_drivers
so that we can list the specific modules that are exclusively needed for
this target. That way we can add new target and not have to touch the
configuration for other targets.

Note that some other dracut modules add modules to the initrd. For
example, systemd adds the ipv6 module. Let's go ahead and exclude
those.

Signed-off-by: Brian Masney <bmasney@redhat.com>