qm: Fix selinux policy
The setup script in upstream qm rpm does:
    rm -rf "${ROOTFS}"/etc/selinux/targeted/contexts/files/file_contexts/*
before installing the custom file_contexts.
This does the same in qm.mpp.yml.

Without this the custom file_contexts will not be used because there
is a pre-compiled file_contexts.bin that will be used instead.

This will help with https://gitlab.com/CentOS/automotive/sample-images/-/issues/19