runosbuild: Fix osbuildvm fallback logic
Commit c8cdda2d5fa5 ("Don't run osbuild with sudo the it runs as root")
changed the fallback logic to osbuildvm.

Before that commit, osbuild was used with sudo if the host and target
architectures were the same, and if the VM variable was set to 0.

osbuildvm was thus called if any of these conditions were false, so if
the host and target architectures were different, or if the VM variable
was set to a non-zero value.

After that commit an additional condition was added to both the initial
osbuild command and the osbuildvm fallback.

Indeed, the logic is a bit more complicated but:

  - sudo osbuild is now run if the host and target architectures are the
    same, if the VM variable is set to 0, and if the script isn't run as
    root;

  - If any of these conditions are false, but only if the VM variable is
    set to 1, we fallback to osbuildvm;

  - If neither sudo osbuild nor osbuildvm can run, we fallback to
    osbuild.

Which means that osbuildvm isn't the default fallback anymore but
explicitly requires the VM variable to be set to 1 now. This is a change
from the previous behavior, so the current commit reinstates the
previous one by moving the sudo osbuild vs osbuild conditions into a
branch of its own, and if neither should run we will fallback to
osbuildvm.

Closes #20.

Signed-off-by: Maxime Ripard <mripard@kernel.org>