Add ccimx93dvk target
Add a target to ease rootfs generation for the Digi Connectcore 93
board.
Note: the developper image barelly fits in the 3GB a/b partition slots.

The bootstrap is still handled manually, so testing this requires
fetching the kernel and initramfs from the rootfs (where they ultimately
do not belong) and copying them to the boot partition for u-boot to
load.

An ext4.simg is required for flashing the rootfs with uuu (fastboot).

Signed-off-by: Eric Chanudet <echanude@redhat.com>