always use lz4-compressed initrd
The initrd appeared to be uncompressed to help with boot speed based on
the configuration, however the initrd that was generated was actually
gzip compressed, and this added about 100ms or so to the overall boot
time.

We can now use lz4 for the rpm-ostree since this merge request has now
been merged: https://github.com/coreos/rpm-ostree/pull/4705

Here's the boot time results now that lz4 is always used:

    # cs9-qemu-minimal-regular.x86_64.qcow2
    Startup finished in 216ms (kernel) + 1.552s (initrd) + 2.448s (userspace) = 4.218s
    Startup finished in 216ms (kernel) + 1.514s (initrd) + 2.156s (userspace) = 3.887s
    Startup finished in 212ms (kernel) + 1.446s (initrd) + 2.180s (userspace) = 3.840s

    # cs9-qemu-minimal-ostree.x86_64.qcow2
    Startup finished in 216ms (kernel) + 1.689s (initrd) + 2.669s (userspace) = 4.575s
    Startup finished in 207ms (kernel) + 1.643s (initrd) + 2.204s (userspace) = 4.056s
    Startup finished in 209ms (kernel) + 1.124s (initrd) + 1.373s (userspace) = 2.707s

Note that this change requires a patch to be applied to ABL for RideSX4
that removes the 'ro' kernel command line argument.

Signed-off-by: Brian Masney <bmasney@redhat.com>