add support for Fedora ELN distro
Add support for Fedora ELN since this is what the next major version of
centos-stream will use as a base.

This change was tested with 'make eln-qemu-minimal-regular.x86_64.qcow2'

    [root@localhost ~]# grep PRETTY_NAME /etc/os-release
    PRETTY_NAME="Fedora ELN"

    [root@localhost ~]# uname -a
    Linux localhost.localdomain 6.8.0-0.rc0.20240118git296455ade1fd.8.eln134.x86_64 #1 SMP PREEMPT_DYNAMIC Thu Jan 18 15:12:02 UTC 2024 x86_64 GNU/Linux

    [root@localhost ~]# cat /etc/redhat-release
    Fedora release 40 (Rawhide)

Note that at the moment that this uses the RHEL kernel since kernel-ark
doesn't have an automotive kernel variant available yet. Once that's
available, we can update the ELN configuration to use the automotive
kernel packages.

Signed-off-by: Brian Masney <bmasney@redhat.com>