Enable transient /etc by default when using composefs
To safely use an immutable system we can't have a persistenly writable
/etc. If /etc is writable and persistent, an attacker could easily
e.g. add a systemd unit that starts evey boot to persist an attack.

To make this impossible this we enable transient_etc which makes /etc
be a writable tmpfs layer over the /etc from the image.

NOTE: This means you now cannot persist changes to /etc over boot,
such changes need to happen each boot, or as part of /etc in the
ostree image.