runvm: amend ovmf firmware location
find_ovmf() returns the first existing directory in the list, even if
the directory in question does not contain the expected firmware
binaries. Fail early if the required OVMF binaries are not available
under the expected path after trying to find the actual binaries under
each known path.

If the --ovmf-dir option is used, the directory passed in shall
take precedence over the default directories in the list.

Signed-off-by: Christophe Lizzi <clizzi@redhat.com>
Signed-off-by: Eric Chanudet <echanude@redhat.com>