Drop use_containers option
Its easy to accidentally not set this, even if containers are used
(for example, its not set in the development manifest), so lets just
always install the custom config files. They are small and unused anyway.

The only issues are that we have to ensure /etc/containers exists, and
that we have to make the enablement of the
podman-clean-transient.service systemd unit optional depending on if
the podman rpm is installed.