Add use_containers_extra_store
This partially adds back something like the use_container option, and
it needs to be set in all images that install containers to the
containers extra store.

This is needed because we can't point the config file to this directory
if it doesn't exist, because podman then breaks trying to create lock
files in this directory, as it is read-only.

I believe this has been fixed in podman upstream (it doesn't fail
anymore), but for now we need this.

Still, this is better then the old use_containers option, as even if
you don't set this you still get other container options like
transient store.