Add new target to generate core rpm lists
This replaces the recent minimal-target with a more generic one.
The basic idea is that if you run e.g:

  make cs9-core-rpms-x86_64.txt

Then this file will be generated that lists the core rpms needed
for the specified distribution configuration (cs9) and arch (x86_64).

The basis for this is the .rpmlist target, built against the
qm-minimal image (a minimal image with qm support) in a variety of
combinations (both ostree and regular) and then the union of all the
rpms used are extracted.