Remove aboot-update from base_rpms
aboot-update is not required at runtime, only when building images, so
we drop it from base_rpms. It is still in build_rpms so it can be used
during image build.

This is important as it drops the python3 dependency from the aboot
target images.

We also add aboot-update to the developer image, as it may be useful
when doing development.