auto-image-builder.sh: fix missing parameters
Using '$@' for passing the 'make' parameters to the container wasn't working.
It was passing just the first parameter, the make target (the image file name),
but not the rest of parameters.
Using '$*' allows to pass the whole array as a string, so it pass as one,
but then be used normally as different parameters.