targets: Ensure cs9-ti-fw has higher priority over other repos
The f39-j784s4evm-developer-regular.aarch64.img build failed during its
copy step:

    copying '/run/osbuild/inputs/rootfs/usr/share/uboot/j784s4_evm_r5/tiboot3-j784s4-gp-evm.bin' -> '/run/osbuild/tree/boot/efi/tiboot3.bin'
    cp: cannot stat '/run/osbuild/inputs/rootfs/usr/share/uboot/j784s4_evm_r5/tiboot3-j784s4-gp-evm.bin': No such file or directory

This is because it used fedora's version of uboot-images
instead of the csi-ti-fw version.

Since that repo definition is intended to mean "take the latest version of
uboot-images from cs9-ti-fw" let's crank the priority to guarantee
that.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>