Drop fstab (and fstab-generator) in favour of mount units
We create mount units similar to the once systemd-fstab-generator
would generate, and then mask out the generator.

On the regular image his makes the systemd generator time drop to almost
zero (although withg ostree there are still some generators in use).

Note: systemd-fstab-generator normally also adds a wanted on
systemd-remount-fs.service. This will remount the rootfs mount as
read-write. This currently only happens in the "regular" target, as
the ostree target specifies "rw" and not "ro" (or it fails). This PR
changes both to use "rw" for consistency and overall performance.