Remove osbuild packages and repos
This has been a pain point for quite a while, our infratructure around
these osbuild repos and rpms has been unstable for quite a while.

We have a Containerized version of the osbuild version of the tool which
is stable and easier to maintain. And there's the python-based
Automotive Image Builder coming also.

Signed-off-by: Eric Curtin <ecurtin@redhat.com>