a-i-b: Automatically quote simple strings
For define values that don't start with a numeric and don't contain
any escapes or typical json type characters we try to auto-quote them
as strings (if the basic json parsing fails).

This allows you do do things like --define foo=bar instead of --define
`foo="bar"`, which is nice on the commandline.