aib.main: fix relative inputs
Since we rewrite the manifest in a new
path at tmpdir, if any of the stages has
a relative path, the build process will fail.

To solve it, walk the entire pipeline and
look for paths. If we encounter a relative
path, make it absolute with respect to the
original manifest's dirname.

Add a test covering all usecases for
rewrite_manifest function.

Signed-off-by: Albert Esteve <aesteve@redhat.com>