add kernel cert and module signing stages
Create a kernel-keys pipeline to hold the signing-key/certificate for
the kernel modules/keyring.

Add org.osbuild-auto.kernel stages to sign the installed modules and
insert an extra certificate to the kernel bootable image. These steps
are conditioned under sign_kernel_modules, which is false by default.

The kernel needs to be configured specifically for it and the stages
only support a subset of the possible kernel image formats and
compression available.

Signed-off-by: Eric Chanudet <echanude@redhat.com>