Split defines from the first '=' character
The defines should take always the portion before the first '='
as the key and the rest as the value.
The current splitting was allowing to split in 3, if a second '='
was present. This could lead to problems when the value contains
that character ('=') for some reason, like a ssh key.