Do not specify start offset for efi partition
By default, 2048 is the starting offset for the first partition, sfdisk
doesn't allow to change this value unless you force it. Even you specify
'0' for gpt type partitions this is totally ignored, and the first
partition is placed at a offset of 2048. So remove the line that says
that the offset is 0 for gpt partition as is not taken into account and
confuses the user.

Signed-off-by: Enric Balletbo i Serra <eballetbo@redhat.com>