Add an optional empty partition for the SBL
Some boards, i.e NXP S32G3 boards, needs some space at the beginning of
the disk to place their SBL (Second Boot Loader). Typically there is
1MB from the start of the disk to the first partition, but, sometimes, this
is not enough because the SBL is bigger. To handle this add a new variable
that allows you to create an empty partition at the beginning of the disk
(4MB by default but configurable) so the boards that requires can leave free
space before the first EFI/BOOT partition to allow their SBL.

Signed-off-by: Enric Balletbo i Serra <eballetbo@redhat.com>