Fix container use when base_dir is in a symlink
If some parent of base_dir is a symlink (for example if you use a-i-b
as a submodule and sample-images is in some symlink) then we can't
start the a-i-b binary via basedir via the symlinks, because the volumees
use realpath(), so they will have different paths.

To fix this we use realpath() on the base_dir too.