sign modules by default with kernel-automotive ostree
JIRA: https://issues.redhat.com/browse/RHEL-53349

Run the step to sign the modules by default if the kernel-automotive
package is selected and mode is "image". That kernel is intended to
support these steps with image mode. The linked ticket will re-introduce
module signing at compilation in order to fix package images that would
want to install an RPM package for testing nonetheless.

In order to make this less intrusive to setup per distro, target, image,
compute the value based on the kernel package and image mode, but should
someone want to force it for some other kernel or to test in package
mode, the parameter is still available.

Signed-off-by: Eric Chanudet <echanude@redhat.com>