Add tuned profile configuration option
This adds an option to enable 'tuned' to be installed and enabled
in the image.
The goal is to have it set the performance tunables and reuse the
parameters already tuned from RHEL.

For some profiles such as realtime, tuned also relies on a set of
kernel parameters at boot. Currently it relies on grub, so that will
not work on images built without it. In order to fetch these parameters
at boot time and allow tuned to be fully used in these situations,
there is currently a feature request upstream:
https://github.com/redhat-performance/tuned/issues/675
Until then, the suggested workaround is to set the kernel parameters
to the ones that tuned would add.

As of this MR, this only targets developer-regular images.

Signed-off-by: Erico Nunes <ernunes@redhat.com>