Restructure include handling of snippets
This changes the structure of includes of the various yaml snippets
with the goal of simplifying maintainance of the various a-i-b files
and making writing manfests more powerful and understandable.

The order of evaluation of the various mpp files when handling
includes is a bit weird. First the main yaml has mpp-vars evalutated,
and then the references includes are loaded and evaluated in include
order. This also applies recursively, further complicating things.

This makes reasoning about evalutation order hard, and additionaly:

1) The manifest file cannot access variables that are defined by
   default or for the chosed arch/target/distro. This is problematic
   for e.g. the qa manifest that want to e.g customize the external /var
   config based on the value of use_aboot.

2) The file that sets the defaults has to manually check for each
   default value if the variable was already set by the user before
   setting it. This leads to some very ugly and hard to maintain code.

The new model is to always pass the "main.ipp.yml" file to
osbuild-mpp.  This then includes the various files (including the
end-user manifest) in a linear order, making the evaluation order
of things very clear, and allowing the end-user manifest to access
the previously set variables.

The new order for variable expansion is:
 * Compute defaults
   * pure default values
   * distro defaults
   * target defaults
   * arch defaults
   * computed defaults (depends on distro/target/arch)
 * Manifest variables
 * Computed variables, as needed for the pipelines

Also, the files with pipelines have been split up and
ordeder in one place such that the order is right allowing
later pipelines to refer to earlier ones.

To get the order of evaluation of the the "name" variable right we
have to do some custom work. This now allows us to pre-initialize it
to the manifest basename.

This looks like a large change, but I ran all the compose-tests and
the produced composed json files were identical to the previous ones
(other that the timestamps, etc).